% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input.R
\name{mod_input_ui}
\alias{mod_input_ui}
\alias{mod_input_server}
\title{mod_input_ui and mod_input_server}
\usage{
mod_input_ui(id)

mod_input_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
